+++ src/components/Profile/ProfilePage/ProfilePageSections/ProfileHeaderContainer/HeaderImage/HeaderImageStyles.tsx
@@ -1,11 +1,11 @@
 import { StyleSheet } from "react-native";
-import metrics from "../../../../../config/metrics";
+import metrics from "../../../../../../config/metrics";
 import {
   white,
   lighterGrey,
   midLightGrey,
   black
-} from "../../../../../themeColors";
+} from "../../../../../../themeColors";
 
 export const styles = StyleSheet.create({
   container: { alignItems: "center" },
