+++ src/components/Profile/ProfilePage/ProfilePageSections/ProfileHeaderContainer/HeaderInformation/NameFin.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { View, Text } from "react-native";
-import { verificationStatusEnum } from "../../../../../services/verificationService/verificationService";
+import { verificationStatusEnum } from "../../../../../../services/verificationService/verificationService";
 import { styles } from "./HeaderInformationStyles";
 
 interface NameFinProps {
