+++ src/components/Profile/ProfilePage/ProfilePageSections/ProfileHeaderContainer/HeaderInformation/index.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from "react";
 import { View, Text, TouchableOpacity } from "react-native";
 import { AntDesign } from "@expo/vector-icons";
-import { SharePageContainer } from "../../../../SharePage";
+import { SharePageContainer } from "../../../../../SharePage";
 import { styles } from "./HeaderInformationStyles";
-import { verificationStatusEnum } from "../../../../../services/verificationService/verificationService";
-import { midGrey } from "../../../../../themeColors";
+import { verificationStatusEnum } from "../../../../../../services/verificationService/verificationService";
+import { midGrey } from "../../../../../../themeColors";
 import { NameFin } from "./NameFin";
-import { profileTypeEnum } from "../../../profileTypeEnum";
+import { profileTypeEnum } from "../../../../profileTypeEnum";
 
 interface HeaderInformationProps {
   status: number;
