+++ src/components/Profile/ProfilePage/ProfilePageSections/ProfileHeaderContainer/ProfileIndicator/index.tsx
